Set `post-self-insert-hook` locally in minor mode
Rather than adding `electric-spacing-post-self-insert-function` to
`post-self-insert-hook` globally when `electric-spacing.el` is loaded we
add it to the buffer-local hook when the minor mode is enabled.

This has the (fairly minor) advantage of not calling
`electric-spacing-post-self-insert-function` unless the minor mode is
enabled (whereas before the function was always called but did nothing
if the minor mode was not enabled).

It also means that merely loading the library doesn't modify any hooks.
I think this is a little bit "nicer".