Properly disable electric spacing in comments/strings
The old implementation doesn't work for me: it still does some odd
things with spaces before the operator. I think the issue was that
`electric-spacing-docs` was only checked inside electric-spacing-insert
but for some cases extra spacing is inserted outside of that
function (e.g. when doulbe spacing after '.').

This implementation does a check in the "top level" function to see if
we are in a comment/string and `electric-spacing-docs` is nil. If so
then nothing is done at all.