Use a hash table for major mode rule lists
This is better than storing each rule list in a variable because it's
extensible.

I used a hash table rather than an alist because alists cannot contain
lists as the value part of the pair (because `(a . b)` where `b` is a
list is exactly the same as `'(a b_0 b_1 ...)`).

It might also be faster to use a hash table.