Fix comment operators in c-mode
Don't add space before a line comment if it is at the indentation of
the current line (single line comment).

Add a space before a line comment if it is not at the indentation of
the current line.

This requires to keep track of the actual insertion position (ignoring
leading white-space) and passing it to the operator. This might also
be useful for enhancing other c-mode operators by querying cc-mode for
the syntactic symbols at the insertion point.