Cache action patterns to reduce memory footprint
Creating the pattern for each operation every time
post-self-insert-function is called results in a lot of unnecessary
memory allocations. As the pattern doesn't change, caching the
generated regex avoids generating it for each insertion.

Typing "a+b-c==1;//This is a test//" allocates ~250KB just for
generating the patterns.

When typing more this increases quickly. Typing 6 include directives
with an average length of 40 characters requires ~20MB.

Before: - electric-operator-post-self-insert-function 44,935,036B  44%
After:  - electric-operator-post-self-insert-function 25,337,486B  26%