Fix unary prefix and prefix in-/decrement operators
When at the beginning of a line, unary prefix and prefix in-/decrement
operators shouldn't be prefixed with a space.

If c-electric-state is on, this will work as the c-electric operations
will fix the indentation. However, if the electric-state is turned off
the additional inserted white-space is not cleaned up and the
indentation was incorrect